refactor(billing): extract utils and style constants to utils.ts
